Propagate ListPropertyField updates to target settings list
